Update Dockerfile
Not sure if there was a reason for the first 'apk add --update curl' line since it is in the big single step block as well... It added an extra docker layer that wasn't needed, so pulled it.  Also, not sure why the Java environment variables were being set when the image doesn't contain any Java, so pulled those out as well.